# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: OmniBid India API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /settings/my:
    get:
      operationId: getMySettings
      tags: [settings]
      summary: Get authenticated user settings
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: User settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettingsResponse"
    put:
      operationId: updateMySettings
      tags: [settings]
      summary: Update authenticated user settings
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettingsResponse"

  # ANALYTICS
...
  /referrals/my:
    get:
      operationId: getMyReferrals
      tags: [referrals]
      summary: Get my referrals and stats
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Referral stats and list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferralStats"
  /referrals/invite:
    post:
      operationId: inviteReferral
      tags: [referrals]
      summary: Send a referral invite
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
              required: [email]
      responses:
        "201":
          description: Referral created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Referral"

  # ENTERPRISE
...
components:
  schemas:
...
    SettingsResponse:
      type: object
      properties:
        settings:
          type: object
          additionalProperties: true
        role:
          type: string
        updatedAt:
          type: string
      required: [settings, role, updatedAt]
...
    ReferralStats:
      type: object
      properties:
        referralCode:
          type: string
        referralLink:
          type: string
        totalReferrals:
          type: number
        convertedReferrals:
          type: number
        totalRewardEarned:
          type: number
        pendingReward:
          type: number
        referrals:
          type: array
          items:
            $ref: "#/components/schemas/Referral"
      required: [referralCode, referralLink, totalReferrals, convertedReferrals, totalRewardEarned, pendingReward, referrals]
    Referral:
      type: object
      properties:
        id:
          type: string
        refereeEmail:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, clicked, signed_up, converted, rewarded]
        rewardAmount:
          type: number
        createdAt:
          type: string
        convertedAt:
          type: string
          nullable: true
      required: [id, status, rewardAmount, createdAt]
...